--- a/.cargo/config.toml
+++ b/.cargo/config.toml
@@ -1,0 +1,3 @@
+
+[build]
+rustflags = ["-C", "symbol-mangling-version=v0"]