--- a/ci/src/commands/clippy.rs
+++ b/ci/src/commands/clippy.rs
@@ -1,6 +1,9 @@
-use crate::{Flag, Prepare, PreparedCommand};
 use argh::FromArgs;
 use xshell::cmd;
+
+use crate::Flag;
+use crate::Prepare;
+use crate::PreparedCommand;
 
 /// Check for clippy warnings and errors.
 #[derive(FromArgs, Default)]