--- a/ci/src/commands/compile_check.rs
+++ b/ci/src/commands/compile_check.rs
@@ -1,6 +1,9 @@
-use crate::{Flag, Prepare, PreparedCommand};
 use argh::FromArgs;
 use xshell::cmd;
+
+use crate::Flag;
+use crate::Prepare;
+use crate::PreparedCommand;
 
 /// Checks that the project compiles.
 #[derive(FromArgs, Default)]