--- a/ci/src/commands/format.rs
+++ b/ci/src/commands/format.rs
@@ -1,6 +1,9 @@
-use crate::{Flag, Prepare, PreparedCommand};
 use argh::FromArgs;
 use xshell::cmd;
+
+use crate::Flag;
+use crate::Prepare;
+use crate::PreparedCommand;
 
 /// Check code formatting.
 #[derive(FromArgs, Default)]