--- a/ci/src/commands/loom_clippy.rs
+++ b/ci/src/commands/loom_clippy.rs
@@ -1,6 +1,9 @@
-use crate::{Flag, Prepare, PreparedCommand};
 use argh::FromArgs;
 use xshell::cmd;
+
+use crate::Flag;
+use crate::Prepare;
+use crate::PreparedCommand;
 
 /// Checks for clippy warnings and errors in the loom test suite.
 #[derive(FromArgs, Default)]