--- a/ci/src/main.rs
+++ b/ci/src/main.rs
@@ -6,7 +6,8 @@ mod ci;
 mod commands;
 mod prepare;
 
-pub use self::{ci::*, prepare::*};
+pub use self::ci::*;
+pub use self::prepare::*;
 
 fn main() {
     argh::from_env::<CI>().run();