--- a/src/lib.rs
+++ b/src/lib.rs
@@ -22,11 +22,16 @@
 //! scheduling. Support for futures is based on an approach sketched out by
 //! members of the `rayon` community to whom we are deeply indebted.
 
+// -----------------------------------------------------------------------------
+// Boilerplate for building without the standard library
+
 #![no_std]
 
+extern crate alloc;
 extern crate std;
 
-extern crate alloc;
+// -----------------------------------------------------------------------------
+// Modules
 
 pub mod job;
 pub mod latch;
@@ -35,10 +40,112 @@ pub mod thread_pool;
 
 mod util;
 
-/// Exports a few commonly used types to simplify working with Forte.
+// -----------------------------------------------------------------------------
+// Prelude
+
 pub mod prelude {
+    //! Reexports some types commonly needed for using Forte.
+
     pub use crate::{
         scope::Scope,
         thread_pool::{ThreadPool, WorkerThread},
     };
+}
+
+// -----------------------------------------------------------------------------
+// Mocked APIs
+
+// This crate uses `loom` for testing, which requires mocking all of the core
+// threading primitives (`Mutex` and the like). Unfortunately there are some
+// minor differences between the `loom` and `std`.
+//
+// To make things a bit simpler, we re-export all the important types in the
+// `primitives` module. Where necessary we wrap the `std` implementation to make
+// it match up with `loom`.
+
+#[cfg(not(loom))]
+mod primitives {
+    pub use core::cell::Cell;
+    pub use core::sync::atomic::AtomicBool;
+    pub use core::sync::atomic::AtomicUsize;
+    pub use core::sync::atomic::Ordering;
+
+    pub use std::sync::Condvar;
+    pub use std::sync::Mutex;
+    pub use std::thread::available_parallelism;
+    pub use std::thread::spawn as spawn_thread;
+
+    pub struct UnsafeCell<T> {
+        data: core::cell::UnsafeCell<T>,
+    }
+
+    impl<T> UnsafeCell<T> {
+        #[inline(always)]
+        pub const fn new(data: T) -> Self {
+            UnsafeCell {
+                data: core::cell::UnsafeCell::new(data),
+            }
+        }
+
+        #[inline(always)]
+        pub fn into_inner(self) -> T {
+            self.data.into_inner()
+        }
+
+        #[inline(always)]
+        pub fn get_mut(&self) -> MutPtr<T> {
+            MutPtr {
+                ptr: self.data.get(),
+            }
+        }
+    }
+
+    pub struct MutPtr<T: ?Sized> {
+        ptr: *mut T,
+    }
+
+    #[allow(clippy::mut_from_ref)]
+    impl<T: ?Sized> MutPtr<T> {
+        /// Dereferences the pointer.
+        ///
+        /// # Safety
+        ///
+        /// This is equivalent to dereferencing a *mut T pointer, so all the
+        /// same safety considerations apply here.
+        ///
+        /// Because the `MutPtr` type can only be created by calling
+        /// `UnsafeCell::get_mut` on a valid `UnsafeCell`, we know the pointer
+        /// will never be null.
+        #[inline(always)]
+        pub unsafe fn deref(&self) -> &mut T {
+            // SAFETY: The safety requirements of this pointer dereference are
+            // identical to those of the function.
+            unsafe { &mut *self.ptr }
+        }
+    }
+
+    pub trait WithMut {
+        fn with_mut<R>(&mut self, f: impl FnOnce(&mut usize) -> R) -> R;
+    }
+
+    impl WithMut for AtomicUsize {
+        #[inline(always)]
+        fn with_mut<R>(&mut self, f: impl FnOnce(&mut usize) -> R) -> R {
+            f(self.get_mut())
+        }
+    }
+}
+
+#[cfg(loom)]
+mod primitives {
+    pub use loom::cell::Cell;
+    pub use loom::cell::UnsafeCell;
+    pub use loom::sync::atomic::AtomicBool;
+    pub use loom::sync::atomic::AtomicUsize;
+    pub use loom::sync::atomic::Ordering;
+    pub use loom::sync::Condvar;
+    pub use loom::sync::Mutex;
+    pub use loom::thread::spawn as spawn_thread;
+
+    pub use std::thread::available_parallelism;
 }