--- a/tests/general.rs
+++ b/tests/general.rs
@@ -1,0 +1,3 @@
+//! General integration tests
+
+#![cfg(not(loom))]