--- a/tests/tests.rs
+++ b/tests/tests.rs
@@ -1,0 +1,4 @@
+//! Entrypoint for all of Forte's tests
+
+mod general;
+mod loom;