--- a/tests/tests.rs
+++ b/tests/tests.rs
@@ -2,3 +2,4 @@
 
 mod general;
 mod loom;
+mod miri;