Adds a the benchmark from `chili`, and a several months worth of incremental optimizations and refactors. On this benchmark, the library is now closer to `chili` on small workloads, and competitive or slightly better than `rayon` across the board.

```
fork_join            fastest       │ slowest       │ median        │ mean          │ samples │ iters
├─ baseline                        │               │               │               │         │
│  ├─ (3, 7)         11.72 ns      │ 14.89 ns      │ 11.8 ns       │ 11.91 ns      │ 100     │ 51200
│  ├─ (4, 15)        25.39 ns      │ 27.34 ns      │ 25.55 ns      │ 25.7 ns       │ 100     │ 25600
│  ├─ (5, 31)        54.53 ns      │ 65.59 ns      │ 55.18 ns      │ 55.35 ns      │ 100     │ 12800
│  ├─ (6, 63)        116 ns        │ 226.7 ns      │ 118 ns        │ 120.4 ns      │ 100     │ 6400
│  ├─ (7, 127)       233.8 ns      │ 391.4 ns      │ 235.2 ns      │ 237.2 ns      │ 100     │ 3200
│  ├─ (8, 255)       470.8 ns      │ 494.2 ns      │ 473.5 ns      │ 475.1 ns      │ 100     │ 1600
│  ├─ (9, 511)       942.1 ns      │ 957.8 ns      │ 947.5 ns      │ 949.8 ns      │ 100     │ 800
│  ├─ (10, 1023)     1.884 µs      │ 1.937 µs      │ 1.895 µs      │ 1.898 µs      │ 100     │ 400
│  ├─ (11, 2047)     3.499 µs      │ 5.041 µs      │ 3.541 µs      │ 3.608 µs      │ 100     │ 100
│  ├─ (12, 4095)     7.04 µs       │ 7.874 µs      │ 7.124 µs      │ 7.134 µs      │ 100     │ 100
│  ├─ (13, 8191)     15.29 µs      │ 27.12 µs      │ 15.49 µs      │ 15.84 µs      │ 100     │ 100
│  ├─ (14, 16383)    28.7 µs       │ 35.16 µs      │ 28.95 µs      │ 29.45 µs      │ 100     │ 100
│  ├─ (15, 32767)    52.66 µs      │ 62.74 µs      │ 55.08 µs      │ 55.76 µs      │ 100     │ 100
│  ├─ (16, 65535)    96.87 µs      │ 108.2 µs      │ 101.5 µs      │ 101 µs        │ 100     │ 100
│  ├─ (17, 131071)   176 µs        │ 214.2 µs      │ 189.4 µs      │ 186.1 µs      │ 100     │ 100
│  ├─ (18, 262143)   353.7 µs      │ 395.2 µs      │ 354.6 µs      │ 362.8 µs      │ 100     │ 100
│  ├─ (19, 524287)   714.1 µs      │ 1.032 ms      │ 758.9 µs      │ 781.3 µs      │ 100     │ 100
│  ╰─ (20, 1048575)  1.518 ms      │ 2.088 ms      │ 1.554 ms      │ 1.588 ms      │ 100     │ 100
├─ chili                           │               │               │               │         │
│  ├─ (3, 7)         40.51 ns      │ 8.374 µs      │ 41.51 ns      │ 126.4 ns      │ 100     │ 100
│  ├─ (4, 15)        40.51 ns      │ 6.249 µs      │ 82.51 ns      │ 136.6 ns      │ 100     │ 100
│  ├─ (5, 31)        95.86 ns      │ 276.8 ns      │ 106.2 ns      │ 116.3 ns      │ 100     │ 3200
│  ├─ (6, 63)        165.5 ns      │ 4.499 µs      │ 207.5 ns      │ 260.7 ns      │ 100     │ 100
│  ├─ (7, 127)       338 ns        │ 746.8 ns      │ 460.4 ns      │ 457.8 ns      │ 100     │ 1600
│  ├─ (8, 255)       614.1 ns      │ 1.02 µs       │ 783.3 ns      │ 784.7 ns      │ 100     │ 800
│  ├─ (9, 511)       1.207 µs      │ 5.707 µs      │ 1.666 µs      │ 1.655 µs      │ 100     │ 100
│  ├─ (10, 1023)     2.374 µs      │ 16.54 µs      │ 3.249 µs      │ 3.247 µs      │ 100     │ 100
│  ├─ (11, 2047)     4.499 µs      │ 12.79 µs      │ 6.208 µs      │ 6.095 µs      │ 100     │ 100
│  ├─ (12, 4095)     9.041 µs      │ 18.7 µs       │ 10.2 µs       │ 11.17 µs      │ 100     │ 100
│  ├─ (13, 8191)     17.29 µs      │ 33.16 µs      │ 23.62 µs      │ 22.9 µs       │ 100     │ 100
│  ├─ (14, 16383)    29.66 µs      │ 54.7 µs       │ 44.1 µs       │ 44.48 µs      │ 100     │ 100
│  ├─ (15, 32767)    48.87 µs      │ 103.4 µs      │ 86.37 µs      │ 81.84 µs      │ 100     │ 100
│  ├─ (16, 65535)    88.87 µs      │ 119.7 µs      │ 105.8 µs      │ 105.6 µs      │ 100     │ 100
│  ├─ (17, 131071)   139.6 µs      │ 244.9 µs      │ 202.6 µs      │ 193.1 µs      │ 100     │ 100
│  ├─ (18, 262143)   209.7 µs      │ 364.3 µs      │ 275.2 µs      │ 270.4 µs      │ 100     │ 100
│  ├─ (19, 524287)   357.2 µs      │ 525.9 µs      │ 428.9 µs      │ 428.7 µs      │ 100     │ 100
│  ╰─ (20, 1048575)  543 µs        │ 748.7 µs      │ 654.8 µs      │ 646.1 µs      │ 100     │ 100
├─ forte                           │               │               │               │         │
│  ├─ (3, 7)         27.83 ns      │ 79.59 ns      │ 38.25 ns      │ 38.33 ns      │ 100     │ 12800
│  ├─ (4, 15)        61.36 ns      │ 344.5 ns      │ 85.45 ns      │ 90.31 ns      │ 100     │ 6400
│  ├─ (5, 31)        129.7 ns      │ 560.7 ns      │ 179.2 ns      │ 196.9 ns      │ 100     │ 3200
│  ├─ (6, 63)        265.1 ns      │ 494.2 ns      │ 361.5 ns      │ 377.8 ns      │ 100     │ 1600
│  ├─ (7, 127)       541 ns        │ 4.207 µs      │ 728.7 ns      │ 820.3 ns      │ 100     │ 400
│  ├─ (8, 255)       1.072 µs      │ 2.051 µs      │ 1.468 µs      │ 1.542 µs      │ 100     │ 400
│  ├─ (9, 511)       2.207 µs      │ 11.83 µs      │ 2.999 µs      │ 3.278 µs      │ 100     │ 100
│  ├─ (10, 1023)     4.374 µs      │ 10.16 µs      │ 6.04 µs       │ 6.437 µs      │ 100     │ 100
│  ├─ (11, 2047)     8.624 µs      │ 17.04 µs      │ 13.1 µs       │ 13.2 µs       │ 100     │ 100
│  ├─ (12, 4095)     12.49 µs      │ 29.33 µs      │ 21.58 µs      │ 21.19 µs      │ 100     │ 100
│  ├─ (13, 8191)     25.04 µs      │ 42.54 µs      │ 35.37 µs      │ 34.87 µs      │ 100     │ 100
│  ├─ (14, 16383)    35.12 µs      │ 77.7 µs       │ 56.81 µs      │ 56.99 µs      │ 100     │ 100
│  ├─ (15, 32767)    71.16 µs      │ 110 µs        │ 93.97 µs      │ 93.28 µs      │ 100     │ 100
│  ├─ (16, 65535)    104.6 µs      │ 159.2 µs      │ 126.6 µs      │ 127.6 µs      │ 100     │ 100
│  ├─ (17, 131071)   140.4 µs      │ 203.6 µs      │ 172.9 µs      │ 172.4 µs      │ 100     │ 100
│  ├─ (18, 262143)   213.6 µs      │ 307 µs        │ 261.9 µs      │ 260.3 µs      │ 100     │ 100
│  ├─ (19, 524287)   353.9 µs      │ 540.7 µs      │ 416.8 µs      │ 424.3 µs      │ 100     │ 100
│  ╰─ (20, 1048575)  623.4 µs      │ 829.8 µs      │ 695.8 µs      │ 709.1 µs      │ 100     │ 100
╰─ rayon                           │               │               │               │         │
   ├─ (3, 7)         6.707 µs      │ 311.6 µs      │ 12.02 µs      │ 20.69 µs      │ 100     │ 100
   ├─ (4, 15)        6.083 µs      │ 76.7 µs       │ 11.99 µs      │ 13.84 µs      │ 100     │ 100
   ├─ (5, 31)        6.041 µs      │ 111.4 µs      │ 13.41 µs      │ 17.72 µs      │ 100     │ 100
   ├─ (6, 63)        5.583 µs      │ 52.66 µs      │ 14.27 µs      │ 16.7 µs       │ 100     │ 100
   ├─ (7, 127)       7.749 µs      │ 40.91 µs      │ 20.6 µs       │ 21.04 µs      │ 100     │ 100
   ├─ (8, 255)       8.499 µs      │ 48.79 µs      │ 24.7 µs       │ 25.28 µs      │ 100     │ 100
   ├─ (9, 511)       13.66 µs      │ 59.58 µs      │ 31.24 µs      │ 31.98 µs      │ 100     │ 100
   ├─ (10, 1023)     15.7 µs       │ 54.74 µs      │ 35.35 µs      │ 36.27 µs      │ 100     │ 100
   ├─ (11, 2047)     18.49 µs      │ 69.16 µs      │ 42.43 µs      │ 42.13 µs      │ 100     │ 100
   ├─ (12, 4095)     32.66 µs      │ 142 µs        │ 50.91 µs      │ 51.9 µs       │ 100     │ 100
   ├─ (13, 8191)     31.41 µs      │ 153.3 µs      │ 59.91 µs      │ 59.85 µs      │ 100     │ 100
   ├─ (14, 16383)    43.62 µs      │ 167.5 µs      │ 71.12 µs      │ 72.16 µs      │ 100     │ 100
   ├─ (15, 32767)    58.95 µs      │ 206.5 µs      │ 89.93 µs      │ 92.93 µs      │ 100     │ 100
   ├─ (16, 65535)    86.62 µs      │ 225.9 µs      │ 118.8 µs      │ 119.3 µs      │ 100     │ 100
   ├─ (17, 131071)   144.7 µs      │ 491 µs        │ 237.6 µs      │ 252.2 µs      │ 100     │ 100
   ├─ (18, 262143)   227.9 µs      │ 594 µs        │ 332 µs        │ 344.6 µs      │ 100     │ 100
   ├─ (19, 524287)   394.2 µs      │ 747.2 µs      │ 488.8 µs      │ 506.9 µs      │ 100     │ 100
   ╰─ (20, 1048575)  729.6 µs      │ 1.094 ms      │ 847.1 µs      │ 859.5 µs      │ 100     │ 100
   ```